+++ src/assets/BIG IDEAS.txt
