+++ src/assets/assignment.txt
