+++ src/controllers/CtrlHomeFeedback.java
@@ -62,7 +62,7 @@ public class CtrlHomeFeedback implements Initializable
     private void backButtonAction()
     {
         try {
-            changedPane = FXMLLoader.load(getClass().getResource("../scenes/HomePage.fxml"));
+            changedPane = FXMLLoader.load(getClass().getResource("src/scenes/HomePage.fxml"));
             setNode(changedPane);
 
         }
