+++ src/controllers/CtrlProfile.java
@@ -305,7 +305,7 @@ public class CtrlProfile implements Initializable
     private void setPageSubscriptions()
     {
         try {
-            changedPane = FXMLLoader.load(getClass().getResource("../scenes/ProfileSubscriptions.fxml"));
+            changedPane = FXMLLoader.load(getClass().getResource("src/scenes/ProfileSubscriptions.fxml"));
             setNode(changedPane);
 
         }
@@ -315,7 +315,7 @@ public class CtrlProfile implements Initializable
     private void setPageTrainers()
     {
         try {
-            changedPane = FXMLLoader.load(getClass().getResource("../scenes/ProfileTrainers.fxml"));
+            changedPane = FXMLLoader.load(getClass().getResource("src/scenes/ProfileTrainers.fxml"));
             setNode(changedPane);
 
         }
@@ -325,7 +325,7 @@ public class CtrlProfile implements Initializable
     private void setPageHalls()
     {
         try {
-            changedPane = FXMLLoader.load(getClass().getResource("../scenes/ProfileHalls.fxml"));
+            changedPane = FXMLLoader.load(getClass().getResource("src/scenes/ProfileHalls.fxml"));
             setNode(changedPane);
 
         }
@@ -335,7 +335,7 @@ public class CtrlProfile implements Initializable
     private void setPageMap()
     {
         try {
-            changedPane = FXMLLoader.load(getClass().getResource("../scenes/ProfileMap.fxml"));
+            changedPane = FXMLLoader.load(getClass().getResource("src/scenes/ProfileMap.fxml"));
             setNode(changedPane);
 
         }
